Build, Analyze, and Refactor LLM Workflows/module3/requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
langchain-chroma>=0.1.0
langchain-text-splitters>=0.0.1
openai>=1.0.0
chromadb>=0.4.0
python-dotenv>=1.0.0
redis>=5.0.0
